[context] Move utility functions from ContextUtil to DefaultContext
