+++ javalin/src/main/java/io/javalin/config/JavalinConfig.kt
@@ -6,8 +6,8 @@
 package io.javalin.config
 
 import io.javalin.Javalin
+import io.javalin.http.MAX_REQUEST_SIZE_KEY
 import io.javalin.http.util.AsyncUtil.ASYNC_EXECUTOR_KEY
-import io.javalin.http.util.ContextUtil.MAX_REQUEST_SIZE_KEY
 import io.javalin.plugin.PluginUtil.attachPlugins
 import io.javalin.plugin.json.JSON_MAPPER_KEY
 import io.javalin.plugin.json.JavalinJackson
