+++ javalin/src/main/java/io/javalin/http/SinglePageHandler.kt
@@ -7,7 +7,6 @@
 package io.javalin.http
 
 import io.javalin.http.staticfiles.Location
-import io.javalin.http.util.ContextUtil.isLocalhost
 import io.javalin.util.Util
 import java.net.URL
